Tela Detalhes Item Carteira

+++ app/src/main/java/com/baltazarstudio/regular/database/IDAO.kt
@@ -1,7 +1,7 @@
 package com.baltazarstudio.regular.database
 
 interface IDAO<T> {
-    fun get(id: Long) : T
+    fun get(id: Int) : T
     fun getTodos() : List<T>
     fun inserir(objeto: T)
     fun alterar(objeto: T)
